+++ packages/utils/src/index.ts
@@ -103,3 +103,5 @@ export function getTimezone(): string {
 export * from "./phone";
 export * from "./locale";
 export * from "./address";
+export * from "./datetime";
+